requirements.txt
h5py
lmdb
numpy
pillow
matplotlib
protobuf==3.19